fixed patientlist sortof

+++ src/Project ICT 4 - Medicine Planner (C#)/MedAgent 0.1/MedAgent 0.1/Patient.cs
@@ -13,9 +13,9 @@ namespace MedAgent_0_1
         public char Sex { get; set; }
         public string Email { get; set; }
         public DateTime Bday { get; set; }
-        public int SSN { get; set; }
+        public long SSN { get; set; }
 
-        public int Telephone { get; set; }
+        public long Telephone { get; set; }
         public string Address { get; set; }
         public string MedicineHistory { get; set; }
         public List<string> Symptoms { get; set; }
